tests/wpt/tests/css/css-text/text-align/text-align-webkit-match-parent.html

<!DOCTYPE html>
<title>text-align: -webkit-match-parent should not be supported</title>
<link rel="help" href="https://drafts.csswg.org/css-text/#valdef-text-align-match-parent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
test(function() {
  assert_false(CSS.supports('text-align: -webkit-match-parent'));
});
</script>
